+++ Core/STM32F103/src/f103_gpio.c
@@ -43,14 +43,14 @@ void F103_GPIO_EnablePortClock(void *port)
  * 把单 bit 引脚掩码转换为引脚编号（0~15）。
  * 例如 GPIO_Pin_13 -> 13。
  */
-uint32_t F103_GPIO_PinIndex(uint16_t pin)
+uint32_t F103_GPIO_PinIndex(uint32_t pin)
 {
 	/* index: 遍历引脚编号。 */
 	uint32_t index;
 
 	for (index = 0U; index < 16U; ++index)
 	{
-		if (pin == (uint16_t)(1U << index))
+		if (pin == (1UL << index))
 		{
 			return index;
 		}
@@ -65,7 +65,7 @@ uint32_t F103_GPIO_PinIndex(uint16_t pin)
  * 2) 开启 GPIO 时钟
  * 3) 配置 CRL/CRH 为 2MHz 推挽输出模式
  */
-void F103_GPIO_InitOutput(void *port, uint16_t pin)
+void F103_GPIO_InitOutput(void *port, uint32_t pin)
 {
 	/* gpioPort: GPIO 寄存器映射地址。 */
 	F103_GPIO_Regs_t *gpioPort;
@@ -103,7 +103,7 @@ void F103_GPIO_InitOutput(void *port, uint16_t pin)
 }
 
 /* GPIO 输入初始化：配置为浮空输入。 */
-void F103_GPIO_InitInput(void *port, uint16_t pin)
+void F103_GPIO_InitInput(void *port, uint32_t pin)
 {
 	/* gpioPort: GPIO 寄存器映射地址。 */
 	F103_GPIO_Regs_t *gpioPort;
@@ -140,7 +140,7 @@ void F103_GPIO_InitInput(void *port, uint16_t pin)
 }
 
 /* GPIO 上拉输入初始化：配置为输入上拉，并默认拉高对应引脚。 */
-void F103_GPIO_InitInputPullUp(void *port, uint16_t pin)
+void F103_GPIO_InitInputPullUp(void *port, uint32_t pin)
 {
 	/* gpioPort: GPIO 寄存器映射地址。 */
 	F103_GPIO_Regs_t *gpioPort;
@@ -179,7 +179,7 @@ void F103_GPIO_InitInputPullUp(void *port, uint16_t pin)
 }
 
 /* GPIO 写电平接口：内部使用 BSRR/BRR 原子置位和复位。 */
-void F103_GPIO_Write(void *port, uint16_t pin, uint8_t level)
+void F103_GPIO_Write(void *port, uint32_t pin, uint8_t level)
 {
 	/* gpioPort: GPIO 寄存器映射地址。 */
 	F103_GPIO_Regs_t *gpioPort;
@@ -203,7 +203,7 @@ void F103_GPIO_Write(void *port, uint16_t pin, uint8_t level)
 }
 
 /* GPIO 读电平接口：读取 IDR 对应位。 */
-uint8_t F103_GPIO_Read(void *port, uint16_t pin)
+uint8_t F103_GPIO_Read(void *port, uint32_t pin)
 {
 	/* gpioPort: GPIO 寄存器映射地址。 */
 	F103_GPIO_Regs_t *gpioPort;
